Large set of changes in main etalonanalysis.py file.
- Add additional global constants for order indices (ALL..., TEST...)

- Fix regression that was still saving sigma and boxhalfwidth peak
fitting parameters in (implicit) units of pixels rather than wavelength

- Rewrite pixel_to_wavelength and wavelength_to_pixel functions to
account for 0d array output weirdness, so they are now self-contained

- Add dedicated Peak property returning the RMS value of the residuals

- Allow for order-level spec_residuals (built up equivalent to the
`spec` array) to be relative to the peak heights rather than absolute.
This is the new defaul bahaviour (normalised)

- Add properties to Orders in order to record the original files that
the `spec` and `wave` data came from, as well as a property to interpret
and return the wls_source, either lfc or thorium

- Allow for not auto-loading wls file for a newly created Spectrum
object. This is to allow for...

- Multiple wls_files per spectrum, with a different set of valid
orderlets or order (indices) that they apply to. This is critical to
apply the thorium WLS and LFC WLS to different orders in a single
Spectrum. Because there is no 'master' WLS file generated with the
relevant WLSes applied per order. The default behaviour is still to look
for the relevant LFC WLS file if none is passed in (and `auto_load_wls`
is True). The main function to call in order to load multiple WLSes is
`Spectrum().load_partial_wls()`

- The `test()` example (run with `python -m polly.etalonanalysis`) has
an example of loading the two relevant WLS files for their corresponding
spectral orders

- A few small teaks to the Spectrum diagnostic plots, both visual and in
handling edge cases arising from some of the above changes